Move delegateRegistrationMiddleware to accountMiddleware
