+++ src/store/middlewares/account.js
@@ -1,9 +1,11 @@
 import { getAccountStatus, getAccount, transactions } from '../../utils/api/account';
-import { accountUpdated } from '../../actions/account';
+import { accountUpdated, accountLoggedIn } from '../../actions/account';
 import { transactionsUpdated } from '../../actions/transactions';
 import { activePeerUpdate } from '../../actions/peers';
 import actionTypes from '../../constants/actions';
 import { fetchAndUpdateForgedBlocks } from '../../actions/forging';
+import { getDelegate } from '../../utils/api/delegate';
+import transactionTypes from '../../constants/transactionTypes';
 
 const updateAccountData = next => (store) => { // eslint-disable-line
   const { peers, account } = store.getState();
@@ -12,10 +14,10 @@ const updateAccountData = next => (store) => { // eslint-disable-line
     if (result.balance !== account.balance) {
       const maxBlockSize = 25;
       transactions(peers.data, account.address, maxBlockSize)
-      .then(response => next(transactionsUpdated({
-        confirmed: response.transactions,
-        count: parseInt(response.count, 10),
-      })));
+        .then(response => next(transactionsUpdated({
+          confirmed: response.transactions,
+          count: parseInt(response.count, 10),
+        })));
       if (account.isDelegate) {
         store.dispatch(fetchAndUpdateForgedBlocks({
           activePeer: peers.data,
@@ -35,12 +37,29 @@ const updateAccountData = next => (store) => { // eslint-disable-line
   });
 };
 
+const delegateRegistration = (store, action) => {
+  const delegateRegistrationTx = action.data.confirmed.filter(
+    transaction => transaction.type === transactionTypes.registerDelegate)[0];
+  const state = store.getState();
+
+  if (delegateRegistrationTx) {
+    getDelegate(state.peers.data, state.account.publicKey)
+      .then((delegateData) => {
+        store.dispatch(accountLoggedIn(Object.assign({},
+          { delegate: delegateData.delegate, isDelegate: true })));
+      });
+  }
+};
+
 const accountMiddleware = store => next => (action) => {
   next(action);
   const update = updateAccountData(next);
   switch (action.type) {
     case actionTypes.metronomeBeat:
       update(store);
+      break;
+    case actionTypes.transactionsUpdated:
+      delegateRegistration(store, action);
       break;
     default: break;
   }
