+++ src/store/middlewares/index.js
@@ -6,7 +6,6 @@ import addedTransactionMiddleware from './addedTransaction';
 import loadingBarMiddleware from './loadingBar';
 import offlineMiddleware from './offline';
 import notificationMiddleware from './notification';
-import delegateRegistrationMiddleware from './delegateRegistration';
 
 export default [
   thunk,
@@ -17,5 +16,4 @@ export default [
   loadingBarMiddleware,
   offlineMiddleware,
   notificationMiddleware,
-  delegateRegistrationMiddleware,
 ];
